+++ src/main/kotlin/blackjack/card/Card.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.card
 
 import java.lang.IllegalArgumentException
