+++ src/main/kotlin/blackjack/card/Suit.kt
@@ -1,4 +1,4 @@
-package blackjack.domain
+package blackjack.card
 
 enum class Suit() {
     SPADE,
