+++ src/test/kotlin/blackjack/CardTest.kt
@@ -1,6 +1,6 @@
 package blackjack
 
-import blackjack.domain.Card
+import blackjack.card.Card
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
