+++ src/test/kotlin/blackjack/DealerTest.kt
@@ -1,8 +1,8 @@
 package blackjack
 
-import blackjack.domain.Card
-import blackjack.domain.Rank
-import blackjack.domain.Suit
+import blackjack.card.Card
+import blackjack.card.Rank
+import blackjack.card.Suit
 import blackjack.domain.participant.Dealer
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
