+++ src/test/kotlin/blackjack/DeckTest.kt
@@ -1,9 +1,9 @@
 package blackjack
 
-import blackjack.domain.Card
-import blackjack.domain.Deck
-import blackjack.domain.Rank
-import blackjack.domain.Suit
+import blackjack.card.Card
+import blackjack.card.Deck
+import blackjack.card.Rank
+import blackjack.card.Suit
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.jupiter.api.Test
